Launch QC app

Visualize a QCmetrics object generated using the get_qcmetrics function.

QCapp(
  qcmetrics,
  background = NULL,
  samplesheet = NULL,
  show_all_relationships = TRUE,
  variables = "all"
)

# S4 method for QCmetrics
QCapp(
  qcmetrics,
  background = NULL,
  samplesheet = NULL,
  show_all_relationships = TRUE,
  variables = "all"
)

Arguments

qcmetrics
A QCmetrics object.

background
Object of class QCmetrics or class QCmetricsBG. These metrics will be shown as background in the relevant plots.

samplesheet
A samplesheet (data.frame), optional argument. By default, the samplesheet stored within the QCmetrics object will be used.

show_all_relationships
Show all relationships in relatedness plots? For large data this can slow down the app considerably given that the number of relations is N^2. If `FALSE`, individual-level data is shown only for relations passing the relatedness threshold, all other relations are summarized into hexbins.

variables
Optional, which variables in samplesheet to use.

Examples

library(minfiData)

#> Loading required package: IlluminaHumanMethylation450kmanifest

## Load example data from minfiData package (6 samples)
library(minfiData)
library(dplyr)
baseDir <- system.file("extdata", package="minfiData")
samplesheet <- read.metharray.sheet(baseDir)

#> [read.metharray.sheet] Found the following CSV files:

#> [1] "/Library/Frameworks/R.framework/Versions/3.6/Resources/library/minfiData/extdata/SampleSheet.csv"

# Make sure that the samplesheet contains a 'Sample_Name'
# column with the sample IDs and a 'Sex' column!
samplesheet <- samplesheet %>%
                   dplyr::mutate(Sample_Name =  paste(samplesheet$Slide,
                                 samplesheet$Array, sep = "_")) %>%
                   dplyr::rename(Sex = sex)

## Get QC metrics
metrics <- get_qcmetrics(samplesheet,
  name = "example minfiData",
  calculate_PCs = TRUE,
  n_probes_PCA = 20000,
  heatmap_variables = c("Array", "Slide"),
  heatmap_variables_fct = c("Array", "Slide"))

#> Creating a QCmetrics object
#> Analysis started at: 2021-09-20 09:30:14

#> Array: 450k
#> -----------------------------

#> Extracting QCmetrics in chunks of 40 samples..

#> Using 144/144 probes for alcohol prediction

#> Using 187/187 probes for smoking prediction

#> Using 435/435 probes for BMI prediction

#> Done!
#> Merging the chunks..

#> Performing IBS..

#> 10 out of 65 probes could be accurately called

#> Calculating relatedness using the DNAm-inferred SNPs

#> Performing array-wide PCA..

#> Performing control probe PCA..

#> Performing association tests between array-wide PCs and technical/biological variables..

#> The following variables have zero variance, and will therefore be excluded from the association tests:

#> 	CD8T

#> 	Neutro

#> Performing association tests between control probe PCs and technical/biological variables..

#> The following variables have zero variance, and will therefore be excluded from the association tests:

#> 	CD8T

#> 	Neutro

#> 
#> Merging the metrics..

#> 
#> Number of outliers using default thresholds:

#> MU: 0 outliers (0%)

#> RG_ratio: 0 outliers (0%)

#> GR_ratio: 0 outliers (0%)

#> OP: 0 outliers (0%)

#> HC: 0 outliers (0%)

#> bscon: 0 outliers (0%)

#> detectionP: 2 outliers (33.33%) [WARNING] High percentage of outliers!

#> beadNr: 0 outliers (0%)

#> sex: 0 outliers (0%)

#> Total: 2 outliers (33.33%) [WARNING] High percentage of outliers!

#> Done!

#> Analysis finished at: 2021-09-20 09:30:48
#> 

if (FALSE) {
QCapp(metrics)
}

# By default the samplesheet used in the get_qcmetrics() function will be used.
# However, users can provide an (updated) samplesheet.
# For example: rename the 'age' column to 'Age' so that it will be recognized and used
# to create a scatterplot between Age and Predicted Age
samplesheet <- samplesheet %>% dplyr::rename(Age = age)
if (FALSE) {
QCapp(metrics, samplesheet = samplesheet)
}