SNPlinker file used in MinE QC

Probes used for inferring genotypes in the MinE data. Probes that contain a SNP (MAF > 0.05 in EUR and AMR population) within the 'G' or 'C' base were selected. Furthermore, probes that are present on both the EPIC and the 450k array were selected. Then the beta2genotype() function from the omicsPrint package was used within the MinE 450k data (6556 samples) to select probes that could be accurately called from these probes (426 probes in total).

data(snplinker_mine)

Format

An object of class "tibble"

Examples

library(minfiData)
library(dplyr)
baseDir <- system.file("extdata", package="minfiData")
samplesheet <- read.metharray.sheet(baseDir)

#> [read.metharray.sheet] Found the following CSV files:

#> [1] "/Library/Frameworks/R.framework/Versions/3.6/Resources/library/minfiData/extdata/SampleSheet.csv"

# Make sure that the samplesheet contains a 'Sample_Name'
# column with the sample IDs and a 'Sex' column!
samplesheet <- samplesheet %>%
                   dplyr::mutate(Sample_Name =  paste(samplesheet$Slide,
                                 samplesheet$Array, sep = "_")) %>%
                   dplyr::rename(Sex = sex)

data(snplinker_mine)
metrics <- get_qcmetrics(samplesheet, snplinker = snplinker_mine)

#> Creating a QCmetrics object
#> Analysis started at: 2021-05-14 15:22:26

#> Array: 450k
#> -----------------------------

#> Extracting QCmetrics in chunks of 40 samples..

#> Using 144/144 probes for alcohol prediction

#> Using 187/187 probes for smoking prediction

#> Using 435/435 probes for BMI prediction

#> Done!
#> Merging the chunks..

#> Performing IBS..

#> 49 out of 426 probes could be accurately called

#> Calculating relatedness using the DNAm-inferred SNPs

#> Performing control probe PCA..

#> Performing association tests between control probe PCs and technical/biological variables..

#> The following variables have zero variance, and will therefore be excluded from the association tests:

#> 	CD8T

#> 	Neutro

#> 
#> Merging the metrics..

#> 
#> Number of outliers using default thresholds:

#> MU: 0 outliers (0%)

#> RG_ratio: 0 outliers (0%)

#> GR_ratio: 0 outliers (0%)

#> OP: 0 outliers (0%)

#> HC: 0 outliers (0%)

#> bscon: 0 outliers (0%)

#> detectionP: 2 outliers (33.33%) [WARNING] High percentage of outliers!

#> beadNr: 0 outliers (0%)

#> sex: 0 outliers (0%)

#> Total: 2 outliers (33.33%) [WARNING] High percentage of outliers!

#> Done!

#> Analysis finished at: 2021-05-14 15:22:59
#>